+++ tests/integration/pusher/get.js
@@ -29,19 +29,19 @@ describe("Pusher", function () {
         )
         .reply(200, "{}")
 
-      pusher.get(
-        {
+      pusher
+        .get({
           path: "/channels",
           params: {
             filter_by_prefix: "presence-",
             info: "user_count,subscription_count",
           },
-        },
-        done
-      )
+        })
+        .then(() => done())
+        .catch(done)
     })
 
-    it("should call back with the result", function (done) {
+    it("should resolve to the response", function (done) {
       var mock = nock("http://api.pusherapp.com")
         .filteringPath(function (path) {
           return path
@@ -53,19 +53,19 @@ describe("Pusher", function () {
         )
         .reply(200, '{"test key": "test value"}')
 
-      pusher.get({ path: "/test", params: {} }, function (
-        error,
-        request,
-        response
-      ) {
-        expect(error).to.be(null)
-        expect(response.statusCode).to.equal(200)
-        expect(response.body).to.equal('{"test key": "test value"}')
-        done()
-      })
+      pusher
+        .get({ path: "/test", params: {} })
+        .then(response => {
+          expect(response.status).to.equal(200)
+          return response.text().then(body => {
+            expect(body).to.equal('{"test key": "test value"}')
+            done()
+          })
+        })
+        .catch(done)
     })
 
-    it("should call back with a RequestError if Pusher responds with 4xx", function (done) {
+    it("should reject with a RequestError if Pusher responds with 4xx", function (done) {
       var mock = nock("http://api.pusherapp.com")
         .filteringPath(function (path) {
           return path
@@ -77,17 +77,13 @@ describe("Pusher", function () {
         )
         .reply(400, "Error")
 
-      pusher.get({ path: "/test", params: {} }, function (
-        error,
-        request,
-        response
-      ) {
+      pusher.get({ path: "/test", params: {} }).catch(error => {
         expect(error).to.be.a(Pusher.RequestError)
         expect(error.message).to.equal("Unexpected status code 400")
         expect(error.url).to.match(
           /^http:\/\/api.pusherapp.com\/apps\/999\/test\?auth_key=111111&auth_timestamp=[0-9]+&auth_version=1\.0&auth_signature=[a-f0-9]+$/
         )
-        expect(error.statusCode).to.equal(400)
+        expect(error.status).to.equal(400)
         expect(error.body).to.equal("Error")
         done()
       })
@@ -113,7 +109,10 @@ describe("Pusher", function () {
         )
         .reply(200, '{"test key": "test value"}')
 
-      pusher.get({ path: "/test", params: {} }, done)
+      pusher
+        .get({ path: "/test", params: {} })
+        .then(() => done())
+        .catch(done)
     })
 
     it("should respect the timeout when specified", function (done) {
@@ -135,23 +134,15 @@ describe("Pusher", function () {
         .delayConnection(101)
         .reply(200)
 
-      pusher.get({ path: "/test", params: {} }, function (
-        error,
-        request,
-        response
-      ) {
-        var expectedError = new Error("ESOCKETTIMEDOUT")
-        expectedError.code = "ESOCKETTIMEDOUT"
-        expectedError.connect = false
-
+      pusher.get({ path: "/test", params: {} }).catch(error => {
         expect(error).to.be.a(Pusher.RequestError)
         expect(error.message).to.equal("Request failed with an error")
-        expect(error.error).to.eql(expectedError)
+        expect(error.error.name).to.eql("AbortError")
         expect(error.url).to.match(
           /^http:\/\/api.pusherapp.com\/apps\/999\/test\?auth_key=111111&auth_timestamp=[0-9]+&auth_version=1\.0&auth_signature=[a-f0-9]+$/
         )
-        expect(error.statusCode).to.equal(null)
-        expect(error.body).to.equal(null)
+        expect(error.status).to.equal(undefined)
+        expect(error.body).to.equal(undefined)
         done()
       })
     })