+++ tests/remote/pusher/keepalive.js
@@ -1,3 +1,5 @@
+var https = require("https")
+
 var expect = require("expect.js")
 
 var Pusher = require("../../../lib/pusher")
@@ -8,22 +10,31 @@ describe("Pusher (integration)", function () {
 
     beforeEach(function () {
       pusher = new Pusher.forURL(process.env.PUSHER_URL, {
-        keepAlive: true,
+        agent: new https.Agent({ keepAlive: true }),
       })
     })
 
+    // TODO this test is useless. It passes if keepAlive is false...
     it("should send 2 sequential requests successfully", function (done) {
-      pusher.get({ path: "/channels" }, function (error, request, response) {
-        expect(error).to.be(null)
-        expect(response.statusCode).to.equal(200)
-        expect(JSON.parse(response.body).channels).to.be.an(Object)
-        pusher.get({ path: "/channels" }, function (error, request, response) {
-          expect(error).to.be(null)
-          expect(response.statusCode).to.equal(200)
-          expect(JSON.parse(response.body).channels).to.be.an(Object)
+      pusher
+        .get({ path: "/channels" })
+        .then(response => {
+          expect(response.status).to.equal(200)
+          return response.json()
+        })
+        .then(body => {
+          expect(body.channels).to.be.an(Object)
+          return pusher.get({ path: "/channels" })
+        })
+        .then(response => {
+          expect(response.status).to.equal(200)
+          return response.json()
+        })
+        .then(body => {
+          expect(body.channels).to.be.an(Object)
           done()
         })
-      })
+        .catch(done)
     })
   })
 })