+++ tests/remote/pusher/proxy.js
@@ -1,6 +1,8 @@
 var expect = require("expect.js")
-var http_proxy = require("../../helpers/http_proxy")
 
+var HttpsProxyAgent = require("https-proxy-agent")
+
+var http_proxy = require("../../helpers/http_proxy")
 var Pusher = require("../../../lib/pusher")
 
 describe("Pusher (integration)", function () {
@@ -14,7 +16,7 @@ describe("Pusher (integration)", function () {
 
     beforeEach(function () {
       pusher = new Pusher.forURL(process.env.PUSHER_URL, {
-        proxy: "http://localhost:8321",
+        agent: new HttpsProxyAgent("http://localhost:8321"),
       })
     })
 
@@ -29,30 +31,34 @@ describe("Pusher (integration)", function () {
     describe("#get", function () {
       it("should go through the proxy", function (done) {
         expect(proxy.requests).to.equal(0)
-        pusher.get({ path: "/channels" }, function (error, request, response) {
-          expect(proxy.requests).to.equal(1)
-          expect(error).to.be(null)
-          expect(response.statusCode).to.equal(200)
-          expect(JSON.parse(response.body).channels).to.be.an(Object)
-          done()
-        })
+        pusher
+          .get({ path: "/channels" })
+          .then(response => {
+            expect(proxy.requests).to.equal(1)
+            expect(response.status).to.equal(200)
+            response.json().then(body => {
+              expect(body.channels).to.be.an(Object)
+              done()
+            })
+          })
+          .catch(done)
       })
     })
 
     describe("#trigger", function () {
       it("should go through the proxy", function (done) {
         expect(proxy.requests).to.equal(0)
-        pusher.trigger("integration", "event", "test", null, function (
-          error,
-          request,
-          response
-        ) {
-          expect(proxy.requests).to.equal(1)
-          expect(error).to.be(null)
-          expect(response.statusCode).to.equal(200)
-          expect(JSON.parse(response.body)).to.eql({})
-          done()
-        })
+        pusher
+          .trigger("integration", "event", "test", null)
+          .then(response => {
+            expect(proxy.requests).to.equal(1)
+            expect(response.status).to.equal(200)
+            response.json().then(body => {
+              expect(body).to.eql({})
+              done()
+            })
+          })
+          .catch(done)
       })
     })
   })