+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+## Unreleased
+
+[BREAKING CHANGE] Methods that previously took callbacks now return promises.
+
+- `trigger`, `triggerBatch`, `get`, and `post` return a promise which resolves
+  to a [`Response`](https://github.com/node-fetch/node-fetch#class-response) (as provided by [node-fetch](https://github.com/node-fetch/node-fetch)),
+  or rejects to a `RequestError`. Of particular note:
+    - `Response` has a `status` where is used to have a `statusCode`.
+    - `RequestError` has a `status` where is used to have a `statusCode` to
+      mirror the above.
+    - `Response` has
+      [`json`](https://github.com/node-fetch/node-fetch#bodyjson),
+      [`text`](https://github.com/node-fetch/node-fetch#bodytext), etc to
+      access the body where before it exposed `body` as a string.
+
+[BREAKING CHANGE] The `Pusher` constructor no longer accepts `proxy` or `keepAlive`, but instead accepts [`agent`](https://nodejs.org/api/https.html#https_class_https_agent).
+
+- To configure a proxy, use [`https://www.npmjs.com/package/https-proxy-agent`](https://www.npmjs.com/package/https-proxy-agent) or similar:
+
+```js
+const pusher = new Pusher.forURL(process.env.PUSHER_URL, {
+  agent: new HttpsProxyAgent("http://localhost:8321"),
+})
+```
+
+- To configure keep alive:
+
+```js
+const pusher = new Pusher.forURL(process.env.PUSHER_URL, {
+  agent: new https.Agent({ keepAlive: true }),
+})
+```
+
 ## 3.0.1 (2020-03-27)
 
 [UPGRADED] development dependencies